requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
pytest==8.3.5
httpx==0.28.1
openpyxl==3.1.5
PyJWT==2.9.0
langgraph==1.1.2
